[avi] remove prefilling since it doesn't cater for empty frames that occur mid-stream (e.g. Xvid with B-frames). Empty frames are only treated as delayed if the flush flag is returned by encoder.

+++ branches/avidemux_2.4_branch/avidemux/ADM_outputs/oplug_avi/op_savecopy.h
@@ -22,7 +22,7 @@
      protected :
 		                                EncoderCopy          *copy;	
 		          virtual uint8_t 	setupVideo( char *name  );
-                          virtual uint8_t 	writeVideoChunk(uint32_t frame );
+                          virtual int 	writeVideoChunk(uint32_t frame );
                                   uint8_t       newFile(void);
 
      public:
@@ -54,7 +54,7 @@ class GenericAviSaveCopyUnpack : public   GenericAviSaveCopy
  {
      protected :
                           virtual uint8_t 	setupVideo( char *name  );
-                          virtual uint8_t 	writeVideoChunk(uint32_t frame );
+                          virtual int 	writeVideoChunk(uint32_t frame );
      public:
   
  };
@@ -66,7 +66,7 @@ class GenericAviSaveCopyPack : public   GenericAviSaveCopy
                           uint32_t              curToggle;
                           uint32_t              time_inc;
                           virtual uint8_t 	setupVideo( char *name  );
-                          virtual uint8_t 	writeVideoChunk(uint32_t frame );
+                          virtual int 	writeVideoChunk(uint32_t frame );
                                   uint8_t       prefetch(uint32_t buffer,uint32_t frame);
      public:
                           virtual	~GenericAviSaveCopyPack();
